+++ example.js
@@ -1,3 +1,8 @@
+/* eslint-env node */
+/* eslint no-console: 0 */
+
+"use strict";
+
 // For RunKit:
 const hyphenopoly = require("hyphenopoly");
 
@@ -5,18 +10,26 @@ const hyphenopoly = require("hyphenopoly");
 // const hyphenopoly = require("./hyphenopoly.module.js");
 
 const hyphenator = hyphenopoly.config({
-    "require": ["de", "en-us"],
-    "hyphen": "•",
     "exceptions": {
         "en-us": "en-han-ces"
-    }
+    },
+    "hyphen": "•",
+    "require": ["de", "en-us"]
 });
 
+/**
+ * Asyncly hyphenate english text
+ * @param {string} text - Words to by hyphenated
+ */
 async function hyphenateEn(text) {
     const hyphenateText = await hyphenator.get("en-us");
     console.log(hyphenateText(text));
 }
 
+/**
+ * Asyncly hyphenate germam text
+ * @param {string} text - Words to by hyphenated
+ */
 async function hyphenateDe(text) {
     const hyphenateText = await hyphenator.get("de");
     console.log(hyphenateText(text));