+++ hyphenEngine.asm.js
@@ -1,5 +1,5 @@
 /**
- * @license hyphenEngine.asm.js 2.8.1-devel - client side hyphenation for webbrowsers
+ * @license hyphenEngine.asm.js 3.0.0 - client side hyphenation for webbrowsers
  * ©2019  Mathias Nater, Zürich (mathiasnater at gmail dot com)
  * https://github.com/mnater/Hyphenopoly
  *
@@ -114,7 +114,6 @@ function asmHyphenEngine(std, ext, heap) {
         var i = 0;
         var last_i = 0;
         var charAti = 0;
-        var mode = 0; //0: init/collect patterns, 1: get patlen
         var plen = 0;
         var count = 0;
         var prevWasDigit = 0;
@@ -125,7 +124,9 @@ function asmHyphenEngine(std, ext, heap) {
         var valueStoreStartIndex = 0;
         var valueStoreCurrentIdx = 0;
         var valueStorePrevIdx = 0;
-         var alphabetlength = 0;
+        var alphabetlength = 0;
+        var first = 0;
+        var second = 0;
         valueStoreStartIndex = (valueStoreOffset + 1) | 0;
         valueStoreCurrentIdx = (valueStoreOffset + 1) | 0;
         valueStorePrevIdx = (valueStoreOffset + 1) | 0;
@@ -135,15 +136,24 @@ function asmHyphenEngine(std, ext, heap) {
         i = hpbPatternsOffset | 0;
         last_i = hpbPatternsOffset + patternsLength | 0;
         while ((i | 0) < (last_i | 0)) {
-            charAti = ui8[i | 0] | 0;
-            if ((charAti | 0) == 58) { //58 === ":"
-                mode = !mode;
+            if ((ui8[i | 0] | 0) == 0) {
+                plen = ui8[(i + 1) | 0] | 0;
+                i = (i + 2) | 0;
             } else {
-                if ((mode | 0) == 1) {
-                    plen = charAti | 0;
-                } else {
-                    count = (count + 1) | 0;
-//add to trie
+                if ((ui8[i | 0] | 0) == 255) {
+                    first = ui8[(i + 1) | 0] | 0;
+                    second = ui8[(i + 2) | 0] | 0;
+                    i = (i + 3) | 0;
+                }
+                while ((count | 0) < (plen | 0)) {
+                    if ((count | 0) == 0) {
+                        charAti = first;
+                    } else if ((count | 0) == 1) {
+                        charAti = second;
+                    } else {
+                        charAti = ui8[i | 0] | 0;
+                        i = (i + 1) | 0;
+                    }
                     if ((charAti | 0) > 11) {
                         if ((prevWasDigit | 0) == 0) {
                             valueStoreCurrentIdx = (valueStoreCurrentIdx + 1) | 0;
@@ -168,21 +178,19 @@ function asmHyphenEngine(std, ext, heap) {
                         valueStoreCurrentIdx = (valueStoreCurrentIdx + 1) | 0;
                         prevWasDigit = 1;
                     }
-                    if ((count | 0) == (plen | 0)) {
-//terminate valueStore and save link to valueStoreStartIndex
-                        ui8[(valueStorePrevIdx + 1) | 0] = 255; //mark end of pattern
-                        i32[(patternTrieOffset + rowStart + rowOffset + 4) >> 2] = (valueStoreStartIndex - valueStoreOffset) | 0;
-//reset indizes
-                        valueStoreStartIndex = (valueStorePrevIdx + 2) | 0;
-                        valueStoreCurrentIdx = valueStoreStartIndex;
-                        count = 0;
-                        rowStart = 0;
-                        nextRowStart = 0;
-                        prevWasDigit = 0;
-                    }
+                    count = (count + 1) | 0;
                 }
+//terminate     valueStore and save link to valueStoreStartIndex
+                ui8[(valueStorePrevIdx + 1) | 0] = 255; //mark end of pattern
+                i32[(patternTrieOffset + rowStart + rowOffset + 4) >> 2] = (valueStoreStartIndex - valueStoreOffset) | 0;
+//reset indi    zes
+                valueStoreStartIndex = (valueStorePrevIdx + 2) | 0;
+                valueStoreCurrentIdx = valueStoreStartIndex;
+                count = 0;
+                rowStart = 0;
+                nextRowStart = 0;
+                prevWasDigit = 0;
             }
-            i = (i + 1) | 0;
         }
         return alphabetlength | 0;
     }