+++ CHANGELOG.md
@@ -1,5 +1,21 @@
 # Version History
 
+## Version 3.0.0 (2019-03-29)
+### Changed
+*   BREAKING CHANGE: implement new hyphenation pattern binary (.hpb) format (issue #61)
+    *   update patterns to new format and include left-/rightmin according to source
+    *   update hyphenEngine.asm/.wasm to consume new pattern format
+    *   implement .hpb-version-check in Hyphenopoly.js and hyphenopoly.module.js
+    *   fix issue #65
+
+### Added
+*   feat: tearDown event (issue #67)
+*   feat: loadError event (issue #59)
+*   feat: hyphenopoly.module is now browserifyable. Added new config option "loader"
+
+### Removed
+*   BREAKING CHANGE: remove `Hyphenopoly.setup.classnames` – use [selectors](https://github.com/mnater/Hyphenopoly/wiki/Global-Hyphenopoly-Object#new-with-version-260-selectors) instead
+
 ## Version 2.8.0 (Feb 28, 2019)
 *   the error event now accepts a `lvl` field ("info"/"warn"/"error") and logs accordingly (issue #56)
 *   add list of supported languages in hyphenopoly.module.js (issue #57)